move rra tests out to seperate file
